Refactor to secret queue
