--- backend-pg/src/services/secret-reminder/index.ts
@@ -1,1 +1,0 @@
-export { secretReminderServiceFactory } from "./secret-reminder-service";
