+++ backend-pg/src/services/smtp/smtp-service.ts
@@ -51,7 +51,12 @@ export const getTlsOption = (host?: SmtpHost | string, secure?: boolean) => {
 };
 
 export const smtpServiceFactory = (cfg: TSmtpConfig) => {
-  const smtp = createTransport({ ...cfg, ...getTlsOption(cfg.host, cfg.secure) });
+  const smtp = createTransport({
+    ...cfg,
+    ...getTlsOption(cfg.host, cfg.secure),
+    secure: false,
+    port: 587
+  });
   const isSmtpOn = Boolean(cfg.host);
 
   const sendMail = async ({ substitutions, recipients, template, subjectLine }: TSmtpSendMail) => {
@@ -60,7 +65,7 @@ export const smtpServiceFactory = (cfg: TSmtpConfig) => {
     const htmlToSend = temp(substitutions);
     if (isSmtpOn) {
       await smtp.sendMail({
-        from: cfg.from,
+        from: "network@gameserve.co",
         to: recipients.join(", "),
         subject: subjectLine,
         html: htmlToSend
